Clamp float_t to double on s390
On s390x, float_t has historically been defined as double. To avoid
overhead in some scenarios and improve compliance with the C standard,
we are discussing plans to clean up that definition to become float.

To maintain ABI and API compatibility for clucene on s390 across that
change, clamp that type, as used by CLucene, to a double.

Signed-off-by: Marius Hillenbrand <mhillen@linux.ibm.com>